requirements.txt
requests
bitrix24-rest
aiohttp
fastapi
uvicorn
pydantic-settings
python-dotenv
pytz